Fix date error

+++ TaskManagerApp/Common/Model/UserAuth.swift
@@ -10,7 +10,7 @@ import Foundation
 struct UserAuth: Codable {
     var idToken: String
     var refreshToken: String
-    var expires: Int = 0
+    var expires: Double = 0.0
     var tokenType: String
     
     enum CodingKeys: String, CodingKey {
